controller and route.ts backend logic
